fix 'format' target
A command returning non-zero is an error in Make. If a directory
doesn't exist, then the shell invoked returns a 1 when using [ ] and it
causes Make to report an error. I simply reversed the check with '!' so
that if a directory is missing, a 0 is returned, otherwise the find
command will run.